Do FDO builds with magic insns.
Added a linker script so all symbols from libtcmalloc.a go to the google_malloc section.
This is fixing the case when FDO decides not to inline some cold subcalls of, say, tc_new,
and they end up in .text.

Also, don't build with list-push-magic for now, turns out FDO builds do use the bt instruction
and we need a new replacement.