Add realistic implementation of SLL_Pop.
This uses the size class cache to store the next head pointer. We use
the same software fallback technique in case of a cache miss. Two new
instructions, shrd and shrx, are used to represent the lookup and update
procedures, respectively.

shrx uses a memory operand on the Pop path to prefetch the next element,
whereas it will use a register operand on the Push path to directly add
the new element to the cache.

A realistic implementation of SLL_Push will be the next commit.